+++ BasicAnalysis/src/main/java/com/blendvision/player/basic/analysis/sample/MainActivity.kt
@@ -1,0 +1,74 @@
+package com.blendvision.player.basic.analysis.sample
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModelProvider
+import com.blendvision.player.basic.analysis.sample.databinding.ActivityMainBinding
+import com.kkstream.playcraft.paas.player.common.ContentType
+import com.kkstream.playcraft.paas.player.common.data.SettingOptionConfig
+
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var viewBinding: ActivityMainBinding
+    private lateinit var mainViewModel: MainViewModel
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        viewBinding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(viewBinding.root)
+
+        mainViewModel = ViewModelProvider(this)[MainViewModel::class.java]
+
+        // Initialize the UniPlayer within ViewModel
+        mainViewModel.initializePlayer(context = this)
+
+        configurePlayer()
+
+    }
+
+    private fun configurePlayer() {
+
+        // Setup the type of control panel
+        viewBinding.kksPlayerServiceView.setupControlPanel(
+            autoKeepScreenOnEnabled = true,
+            defaultContentType = ContentType.EMBEDDED
+        )
+
+        // Hide useless options
+        viewBinding.kksPlayerServiceView.configureSettingOption(
+            SettingOptionConfig(
+                forceHideAutoPlay = true
+            )
+        )
+
+        viewBinding.kksPlayerServiceView.setUnifiedPlayer(mainViewModel.getPlayer())
+
+    }
+
+    override fun onStart() {
+        super.onStart()
+        mainViewModel.loadPlayer()
+        mainViewModel.startPlayer()
+    }
+
+    override fun onResume() {
+        super.onResume()
+        mainViewModel.playPlayer()
+    }
+
+    override fun onPause() {
+        super.onPause()
+        mainViewModel.pausePlayer()
+    }
+
+    override fun onStop() {
+        super.onStop()
+        mainViewModel.stopPlayer()
+    }
+
+    override fun onDestroy() {
+        mainViewModel.releasePlayer()
+        super.onDestroy()
+    }
+
+}
