--- BasicCasting/src/androidTest/java/com/kkstream/basiccasting/sample/ExampleInstrumentedTest.kt
@@ -1,22 +1,0 @@
-package com.kkstream.basiccasting.sample
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert.*
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.kkstream.basiccasting.sample", appContext.packageName)
-    }
-}
