+++ BasicCasting/src/main/java/com/blendvision/player/basic/casting/sample/cast_button/CastButton.kt
@@ -1,0 +1,7 @@
+package com.blendvision.player.basic.casting.sample.cast_button
+
+import android.view.View
+
+interface CastButton {
+    fun provideCastButton(): View
+}
