--- BasicCasting/src/main/java/com/kkstream/basiccasting/sample/main/cast_button/CastButton.kt
@@ -1,7 +1,0 @@
-package com.kkstream.basiccasting.sample.main.cast_button
-
-import android.view.View
-
-interface CastButton {
-    fun provideCastButton(): View
-}
