+++ BasicPlayback/build.gradle
@@ -1,7 +1,6 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
-    id 'kotlin-kapt'
 }
 
 android {
@@ -10,9 +9,9 @@ android {
     buildFeatures {
         viewBinding = true
     }
-    namespace 'com.kkstream.uniplayersample'
+    namespace 'com.blendvision.player.basic.playback.sample'
     defaultConfig {
-        applicationId "com.kkstream.uniplayersample"
+        applicationId "com.blendvision.player.basic.playback.sample"
         minSdk BuildVersion.minSdk
         targetSdk BuildVersion.targetSdk
         versionCode 1
@@ -79,6 +78,7 @@ dependencies {
     implementation Libs.IMA
 
 
+    implementation Libs.constraintlayout
     implementation Libs.core_ktx
     implementation Libs.appcompat
     implementation Libs.material
