+++ BasicPlayback/src/main/AndroidManifest.xml
@@ -11,7 +11,8 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.UniPlayerSample">
         <activity
-            android:name=".main.MainActivity"
+            android:name="com.blendvision.player.basic.playback.sample.MainActivity"
+            android:configChanges="keyboardHidden|orientation|screenSize|smallestScreenSize|screenLayout"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
