+++ BasicPlayback/src/main/java/com/blendvision/player/basic/playback/sample/MainViewModel.kt
@@ -1,0 +1,76 @@
+package com.blendvision.player.basic.playback.sample
+
+import android.content.Context
+import androidx.lifecycle.ViewModel
+import com.kkstream.playcraft.paas.player.common.PlayerConfig
+import com.kkstream.playcraft.paas.player.common.UniPlayer
+import com.kkstream.playcraft.paas.player.common.callback.PlayLogger
+import com.kkstream.playcraft.paas.player.common.data.MediaConfig
+
+class MainViewModel : ViewModel() {
+
+    private lateinit var player: UniPlayer
+
+    private val mediaConfig = MediaConfig(
+        source = listOf(
+            MediaConfig.Source(
+                url = "MPD_URL",
+                protocol = MediaConfig.Protocol.DASH
+            )
+        ),
+        title = "CONTENT_TITLE",
+        imageUrl = "COVER_IMAGE",
+        thumbnailSeekingUrl = null,
+        playWhenReady = true,
+        sharedUrl = "SHARED_INFO",
+        description = "DESC_INFO"
+    )
+
+    // Initialize the UniPlayer
+    fun initializePlayer(
+        context: Context
+    ) {
+        player = UniPlayer.Builder(
+            context = context,
+            playerConfig = PlayerConfig(
+                playLogger = object : PlayLogger {
+                    override fun logEvent(eventName: String, properties: Map<String, Any>) {
+
+                    }
+                }
+            )
+        ).build()
+    }
+
+    // Load and start the UniPlayer
+    fun loadPlayer() {
+        player.load(mediaConfig)
+    }
+
+    fun startPlayer() {
+        player.start()
+    }
+
+    // Pause the UniPlayer
+    fun pausePlayer() {
+        player.pause()
+    }
+
+    fun playPlayer() {
+        player.play()
+    }
+
+    // Stop the UniPlayer
+    fun stopPlayer() {
+        player.stop()
+    }
+
+    // Release the UniPlayer resources
+    fun releasePlayer() {
+        player.release()
+    }
+
+    fun getPlayer(): UniPlayer {
+        return player
+    }
+}
